Securite pour Fichier, ASsigne, Contact, Discussion, DiscussionMember

+++ src/main/java/ca/usherbrooke/fgen/api/service/AssigneeService.java
@@ -1,9 +1,15 @@
 package ca.usherbrooke.fgen.api.service;
 
+import ca.usherbrooke.fgen.api.business.Equipe;
+import ca.usherbrooke.fgen.api.business.Tache;
 import ca.usherbrooke.fgen.api.mapper.AssigneeMapper;
+import ca.usherbrooke.fgen.api.mapper.EquipeMapper;
+import ca.usherbrooke.fgen.api.mapper.TacheMapper;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+import org.eclipse.microprofile.jwt.JsonWebToken;
 
 import java.util.List;
 
@@ -15,6 +21,15 @@ public class AssigneeService {
     @Inject
     AssigneeMapper assigneeMapper;
 
+    @Inject
+    TacheMapper tacheMapper;
+
+    @Inject
+    EquipeMapper equipeMapper;
+
+    @Inject
+    JsonWebToken jwt;
+
     @GET
     @Path("/{tacheId}")
     public List<String> getAssignees(@PathParam("tacheId") String tacheId) {
@@ -24,6 +39,16 @@ public class AssigneeService {
     @POST
     @Path("/{tacheId}")
     public String insertAssignee(@PathParam("tacheId") String tacheId, @QueryParam("cip") String cip) {
+        String cipConnecte = jwt.getClaim("cipConnecte");
+
+        Tache tache = tacheMapper.selectOne(tacheId);
+        String equipeId = tache.equipeId;
+        Equipe equipe = equipeMapper.selectOne(equipeId);
+
+        if(equipe == null || !equipeMapper.selectMembers(equipeId).contains(cipConnecte)) {
+            throw new WebApplicationException(Response.Status.FORBIDDEN);
+        }
+
         assigneeMapper.insertAssignee(tacheId, cip);
         return cip;
     }
@@ -31,6 +56,16 @@ public class AssigneeService {
     @DELETE
     @Path("/{tacheId}")
     public String deleteAssignee(@PathParam("tacheId") String tacheId, @QueryParam("cip") String cip) {
+        String cipConnecte = jwt.getClaim("cipConnecte");
+
+        Tache tache = tacheMapper.selectOne(tacheId);
+        String equipeId = tache.equipeId;
+        Equipe equipe = equipeMapper.selectOne(equipeId);
+
+        if(equipe == null || !equipeMapper.selectMembers(equipeId).contains(cipConnecte)) {
+            throw new WebApplicationException(Response.Status.FORBIDDEN);
+        }
+
         assigneeMapper.deleteAssignee(tacheId, cip);
         return cip;
     }
