+++ src/main/java/ca/usherbrooke/fgen/api/service/ContactService.java
@@ -4,6 +4,8 @@ import ca.usherbrooke.fgen.api.mapper.ContactMapper;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+import org.eclipse.microprofile.jwt.JsonWebToken;
 
 import java.util.List;
 
@@ -15,6 +17,8 @@ public class ContactService {
     @Inject
     ContactMapper contactMapper;
 
+    @Inject
+    JsonWebToken jwt;
     @GET
     public List<String> getContacts(@QueryParam("cip") String cip) {
         return contactMapper.selectContacts(cip);
@@ -22,6 +26,12 @@ public class ContactService {
 
     @DELETE
     public String deleteContact(@QueryParam("cip") String cip, @QueryParam("cipContact") String cipContact) {
+        String cipConnecte = (String) jwt.getClaim("cip");
+
+        if(!(cip.equals(cipConnecte) || cipContact.equals(cipConnecte))) {
+            throw new WebApplicationException(Response.Status.FORBIDDEN);
+        }
+
         contactMapper.deleteContact(cip, cipContact);
         contactMapper.deleteContact(cipContact, cip); // bidirectionnel
         return cipContact;
