+++ src/main/java/ca/usherbrooke/fgen/api/service/DiscussionMemberService.java
@@ -1,12 +1,15 @@
 package ca.usherbrooke.fgen.api.service;
 
+import ca.usherbrooke.fgen.api.business.Discussion;
 import ca.usherbrooke.fgen.api.business.DiscussionMemberSummary;
 import ca.usherbrooke.fgen.api.mapper.DiscussionMapper;
 import ca.usherbrooke.fgen.api.mapper.DiscussionMemberMapper;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
 import org.apache.ibatis.annotations.Param;
+import org.eclipse.microprofile.jwt.JsonWebToken;
 
 import java.util.List;
 
@@ -19,6 +22,8 @@ public class DiscussionMemberService {
     DiscussionMemberMapper discussionMemberMapper;
     @Inject
     DiscussionMapper discussionMapper;
+    @Inject
+    JsonWebToken jwt;
 
     @GET
     @Path("/conversations")
@@ -29,12 +34,26 @@ public class DiscussionMemberService {
     @POST
     @Path("/{discussionId}")
     public String changeState(@PathParam("discussionId") String discussionId, @QueryParam("cip") String cip, @QueryParam("etat") String etat) {
+        String cipConnecte = (String)jwt.getClaim("cip");
+        Discussion discussion = discussionMapper.selectOne(discussionId);
+
+        if(discussion == null || !discussion.members.contains(cipConnecte)) {
+            throw new WebApplicationException(Response.Status.FORBIDDEN);
+        }
+
         discussionMemberMapper.changeState(discussionId, cip, etat);
         return "OK";
     }
 
     @DELETE
     public String deleteMember(@Param("discussionId") String discussionId, @QueryParam("cip") String cip) {
+        String cipConnecte = (String)jwt.getClaim("cip");
+        Discussion discussion = discussionMapper.selectOne(discussionId);
+
+        if(discussion == null || discussion.members.contains(cipConnecte)) {
+            throw new WebApplicationException(Response.Status.FORBIDDEN);
+        }
+
         discussionMemberMapper.deleteMember(discussionId, cip);
 
         if (discussionMapper.select(null, null, discussionId) == null) {
